+++ arcane_jaspr_codex/arcane_codex_web/lib/arcane_codex_web.dart
@@ -1,12 +1,8 @@
 /// arcane_codex_web - Main library export
 library arcane_codex_web;
 
-// Export layouts
-export 'layouts/arcane_docs_layout.dart';
-
 // Export components
 export 'components/docs_header.dart';
-export 'components/docs_sidebar.dart';
 export 'components/code_block.dart';
 export 'components/component_demo.dart';
 export 'components/props_table.dart';